+++ resources/js/core/events/index.ts
@@ -1,0 +1,5 @@
+/**
+ * Events Module - Event bus for component communication
+ */
+
+export * from './EventBus';