+++ resources/views/demo/sections/event-bus.blade.php
@@ -1,0 +1,9 @@
+@props(['framework' => 'vanilla', 'prefix' => 'a'])
+
+@php
+    app('accelade')->setFramework($framework);
+@endphp
+
+<x-accelade::layouts.demo-sidebar :framework="$framework" section="event-bus">
+    @include('accelade::demo.partials._event-bus', ['prefix' => $prefix])
+</x-accelade::layouts.demo-sidebar>